修复x86架构打补丁位置向前跳转时目标进程崩溃问题

+++ upatch/upatch-manage/arch/x86_64/ptrace.c
@@ -141,5 +141,5 @@ unsigned long get_new_insn(struct object_file *obj, unsigned long old_addr,
 	*(unsigned int *)(jmp_insn + 1) =
 		(unsigned int)(new_addr - old_addr - 5);
 
-	return *(unsigned int *)jmp_insn;
+	return *(unsigned long *)jmp_insn;
 }
